+++ src/components/Layouts/FloatingActionButton/index.tsx
@@ -11,7 +11,7 @@ import { FC, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 import { useRecoilValue } from 'recoil'
 
-import { CreateFolderDialog } from '@/features/folder/components/CreateFolderDialog'
+import { CreateFolderDialog } from '@/features/folder/components/Dialog/CreateFolderDialog'
 import { useMedia } from '@/hooks/useMedia'
 import { isAuthenticatedState } from '@/states/AuthAtom'
