+++ src/features/folder/routes/FolderDetails.tsx
@@ -16,7 +16,7 @@ import { LinkButton } from '@/components/Elements/Button/LinkButton'
 import { PageHeading } from '@/components/Elements/Heading/PageHeading'
 import { PageLoading } from '@/components/Layouts/PageLoading'
 import { FavoriteFolderButton } from '@/features/favoriteFolder/components/FavoriteFolderButton'
-import { DeleteFolderDialog } from '@/features/folder/components/DeleteFolderDialog'
+import { DeleteFolderDialog } from '@/features/folder/components/Dialog/DeleteFolderDialog'
 import { FolderLinkButton } from '@/features/folder/components/FolderLinkButton'
 import { useFetchFolder } from '@/features/folder/hooks/useFetchFolder'
 import { Link } from '@/features/link/types/Link'
